+++ src/sim_framework/sim_app_base.hpp
@@ -16,14 +16,14 @@ public:
     virtual void step() = 0;
     
     void initialize(DataBusType& data_bus);
-    void check_step(const uint32_t& sim_time_usec);
+    void check_step(const uint64_t& sim_time_usec);
 
     int priority;
     double app_dt_sec;
     
 private:   
-    uint32_t app_dt_usec;
-    uint32_t next_run_time_usec;    
+    uint64_t app_dt_usec;
+    uint64_t next_run_time_usec;    
     bool time_to_step;
     std::string config_path;
